requirements.txt
pydantic
openpyxl
openai
pillow
pytesseract
fastapi
uvicorn
python-multipart
